+++ src/Toml.hs
@@ -8,8 +8,10 @@ import qualified Toml
 
 module Toml
     ( module Toml.Parser
+    , module Toml.Printer
     , module Toml.Type
     ) where
 
 import Toml.Parser
+import Toml.Printer
 import Toml.Type