+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The change log is available [on GitHub][2].
 
 0.0.0
 =====
+* [#3](https://github.com/kowainik/tomland/issues/3):
+  Implement basic TOML parser with `megaparsec`.
+* [#7](https://github.com/kowainik/tomland/issues/7):
+  Implement type safe version of `Value` type as GADT.
+* [#4](https://github.com/kowainik/tomland/issues/4):
+  Implement basic pretty-printer.
 * [#1](https://github.com/kowainik/tomland/issues/1):
   Implement types representing TOML configuration.
 * Initially created.